Moved Form level error message out of FormHeader
