+++ packages/web-forms/src/components/OdkWebForm.vue
@@ -2,7 +2,8 @@
 import { initializeForm, type RootNode } from '@getodk/xforms-engine';
 import Button from 'primevue/button';
 import Card from 'primevue/card';
-import { provide, reactive, ref } from 'vue';
+import PrimeMessage from 'primevue/message';
+import { computed, provide, reactive, ref } from 'vue';
 import FormHeader from './FormHeader.vue';
 
 import QuestionList from './QuestionList.vue';
@@ -34,11 +35,30 @@ const handleSubmit = () => {
 }
 
 provide('submitPressed', submitPressed);
+
+const formErrorMessage = computed(() => {
+	const violationLength = odkForm.value!.validationState.violations.length;
+
+	if(violationLength === 0) return '';
+	else if(violationLength === 1) return '1 question with error';
+	else return `${violationLength} questions with errors`;
+});
+
+const scrollToFirstInvalidQuestion = () => {
+	document.getElementById(odkForm.value!.validationState.violations[0].nodeId + '_container')?.scrollIntoView({
+		behavior: 'smooth'
+	});
+}
 </script>
 
 <template>
 	<div v-if="odkForm" class="odk-form" :class="{ 'submit-pressed': submitPressed }">
 		<div class="form-wrapper">
+			<PrimeMessage v-if="formErrorMessage" v-show="submitPressed" severity="error" icon="icon-error_outline" class="form-error-message" :closable="false">
+				{{ formErrorMessage }}
+				<span class="fix-errors" @click="scrollToFirstInvalidQuestion()">Fix errors</span>
+			</PrimeMessage>
+
 			<FormHeader :form="odkForm" />
 
 			<Card class="questions-card">
@@ -67,6 +87,8 @@ provide('submitPressed', submitPressed);
 	color: var(--text-color);
 
 	.form-wrapper {
+		display: flex;
+		flex-direction: column;
 		max-width: 900px;
 		margin: auto;
 		padding-top: 10px;
@@ -83,6 +105,35 @@ provide('submitPressed', submitPressed);
 			}
 		}
 
+		.form-error-message.p-message.p-message-error {
+			border-radius: 10px;
+			background-color: var(--error-bg-color);
+			border: 1px solid var(--error-text-color);
+			width: 70%;
+			margin: 0rem auto 1rem auto;
+			position: sticky;
+			top: 0;
+			// Some PrimeVue components use z-index.
+			// Default value for those are either 1000 or 1100
+			// So 5000 here is safe.
+			z-index: 5000;
+
+			:deep(.p-message-wrapper) {
+				padding: 0.75rem 0.75rem;
+				flex-grow: 1;
+			}
+
+			:deep(.p-message-text){
+				font-weight: 400;
+				flex-grow: 1;
+
+				.fix-errors {
+					float: right;
+					cursor: pointer;
+				}
+			}
+
+		}
 	}
 
 	.print-button.p-button {
@@ -106,13 +157,25 @@ provide('submitPressed', submitPressed);
 			max-width: unset;
 			padding-top: unset;
 
+			:deep(.title-bar){
+				order: 1;
+			}
+
+			.form-error-message.p-message.p-message-error {
+				margin-top: 1rem;
+				margin-bottom: 0;
+				order: 2;
+			}
+
 			.questions-card {
 				border-radius: unset;
 				box-shadow: unset;
 				margin-top: 0;
+				order: 3;
 			}
+			
 			.footer {
-
+				order: 4;
 				button {
 					margin-right: 20px;
 				}
